+++ r8s/r8scli/group/application.py
@@ -3,6 +3,7 @@ import click
 from r8scli.group import cli_response, ViewCommand
 from r8scli.group.application_licenses import licenses
 from r8scli.group.application_dojo import dojo
+from r8scli.group.application_rabbitmq import rabbitmq
 from r8scli.service.constants import ALLOWED_PROTOCOLS, PROTOCOL_HTTPS
 
 
@@ -129,3 +130,4 @@ def delete(application_id, force):
 
 application.add_command(licenses)
 application.add_command(dojo)
+application.add_command(rabbitmq)